refactor(app): clean up ODD running protocol views
