+++ app/src/organisms/ODD/RunningProtocol/CancelingRunModal/__tests__/CancelingRunModal.test.tsx
@@ -4,7 +4,7 @@ import { describe, it } from 'vitest'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 
-import { CancelingRunModal } from '../CancelingRunModal'
+import { CancelingRunModal } from '..'
 
 const render = () => {
   return renderWithProviders(<CancelingRunModal />, {
