+++ app/src/organisms/ODD/RunningProtocol/index.ts
@@ -3,3 +3,4 @@ export * from './CurrentRunningProtocolCommand'
 export * from './RunFailedModal'
 export * from './RunningProtocolCommandList'
 export * from './RunningProtocolSkeleton'
+export * from './CancelingRunModal'
