+++ app/src/organisms/ODD/RunningProtocol/shared/StopButton.tsx
