+++ app/src/pages/ODD/RunningProtocol/runningprotocol.module.css
@@ -1,0 +1,49 @@
+.container {
+  position: relative;
+  z-index: 0;
+  display: flex;
+  overflow: hidden;
+  flex-direction: column;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  padding: 1.75rem var(--spacing-40) var(--spacing-40);
+}
+
+.navigation_dots {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  margin-top: var(--spacing-32);
+  gap: var(--spacing-16);
+}
+
+.bullet {
+  z-index: 2;
+  width: 0.5rem;
+  height: 0.5rem;
+  border-radius: 50%;
+  transition: all 0.2s ease;
+}
+
+.bullet_active {
+  background: var(--grey-50);
+  transform: scale(2);
+}
+
+.bullet_inactive {
+  background: var(--grey-40);
+  transform: scale(1);
+}
+
+.gradient_overlay {
+  position: absolute;
+  width: 59rem;
+  height: 20.25rem;
+  align-self: flex-end;
+  margin-top: 9.25rem;
+  background: linear-gradient(rgba(255, 0, 0, 0) 85%, #fff);
+}
